Moved html utilities to utils.
