+++ app/utils.py
@@ -1,14 +1,26 @@
 import os
 from datetime import datetime, timedelta
+from typing import Dict, Union
 from uuid import uuid4
 
-from flask import url_for
+from flask import current_app, url_for
 from notifications_utils.template import SMSMessageTemplate, WithSubjectTemplate, get_html_email_body
 from notifications_utils.url_safe_token import generate_token
 import pytz
 from sqlalchemy import func
 
-from app.constants import EMAIL_TYPE, LETTER_TYPE, PRECOMPILED_LETTER, PUSH_TYPE, SMS_TYPE, UPLOAD_DOCUMENT
+from app.constants import (
+    BRANDING_BOTH,
+    BRANDING_ORG_BANNER,
+    EMAIL_TYPE,
+    LETTER_TYPE,
+    PRECOMPILED_LETTER,
+    PUSH_TYPE,
+    SMS_TYPE,
+    UPLOAD_DOCUMENT,
+)
+from app.feature_flags import is_gapixel_enabled
+from app.googleanalytics.pixels import build_dynamic_ga4_pixel_tracking_url
 
 local_timezone = pytz.timezone(os.getenv('TIMEZONE', 'America/New_York'))
 
@@ -153,3 +165,72 @@ def update_dct_to_str(update_dct):
 
 def create_uuid() -> str:
     return str(uuid4())
+
+
+def get_logo_url(base_url, logo_file):
+    """
+    Generate the URL for a logo file based on the configured asset bucket and domain.
+
+    Args:
+        base_url: Base URL (not used but kept for compatibility)
+        logo_file: Filename of the logo
+
+    Returns:
+        str: The fully qualified URL to the logo
+    """
+    bucket = current_app.config['ASSET_UPLOAD_BUCKET_NAME']
+    domain = current_app.config['ASSET_DOMAIN']
+    return f'https://{bucket}.{domain}/{logo_file}'
+
+
+def get_html_email_options(template) -> Dict[str, Union[str, bool]]:
+    """
+    Generate HTML email options dictionary for email rendering.
+
+    This function creates a dictionary of options that will be used when rendering HTML emails.
+    It determines which branding elements to include based on the service's email_branding configuration.
+    If email_branding is None, it will use the default banner.
+    Otherwise, it will include branding elements like colors, logos, and text from the service's email_branding.
+
+    If Google Analytics pixel tracking is enabled, it also adds a tracking URL to the options.
+
+    Args:
+        template: The template object that contains a reference to the service
+                with branding configuration
+
+    Returns:
+        Dict[str, Union[str, bool]]: A dictionary containing HTML email options including:
+            - default_banner: Whether to use the default banner
+            - brand_banner: Whether to use the custom brand banner
+            - brand_colour: The brand color (if available)
+            - brand_logo: URL to the brand logo (if available)
+            - brand_text: The brand text (if available)
+            - brand_name: The brand name (if available)
+            - ga4_open_email_event_url: Google Analytics tracking URL (if enabled)
+    """
+    options_dict = {}
+    if is_gapixel_enabled(current_app):
+        options_dict['ga4_open_email_event_url'] = build_dynamic_ga4_pixel_tracking_url('xx_notification_id_xx')
+
+    service = template.service
+    if service.email_branding is None:
+        options_dict.update({'default_banner': True, 'brand_banner': False})
+    else:
+        logo_url = (
+            get_logo_url(current_app.config['ADMIN_BASE_URL'], service.email_branding.logo)
+            if service.email_branding.logo
+            else None
+        )
+
+        options_dict.update(
+            {
+                'default_banner': service.email_branding.brand_type == BRANDING_BOTH,
+                'brand_banner': service.email_branding.brand_type == BRANDING_ORG_BANNER,
+                'brand_colour': service.email_branding.colour,
+                'brand_logo': logo_url,
+                'brand_text': service.email_branding.text,
+                'brand_name': service.email_branding.name,
+            }
+        )
+
+    return options_dict
